+++ src/inspect_ai/agent/_bridge/model_proxy.py
